+++ pyogrio/geopandas.py
@@ -167,6 +167,9 @@ def read_dataframe(
     if use_arrow:
         meta, table = result
         df = table.to_pandas()
+        if fid_as_index:
+            df = df.set_index(meta["fid_column"])
+            df.index.names = ["fid"]
         geometry_name = meta["geometry_name"] or "wkb_geometry"
         if geometry_name in df.columns:
             df["geometry"] = from_wkb(df.pop(geometry_name), crs=meta["crs"])