+++ CHANGES.md
@@ -5,6 +5,8 @@
 ### Bug fixes
 
 -   Fix int32 overflow when reading int64 columns (#260)
+-   Fix `fid_as_index=True` doesn't set fid as index using `read_dataframe` with
+    `use_arrow=True` (#265)
 
 ## 0.6.0 (2023-04-27)